Error located
Error wasn't run time. It was a programmger's error. When user gets
points on first roll. User then will have to get the same point to win.
However, if user gets a 7 before his point. User loses. However, I
forgot to put a prompt that user loses and a character catch to stop the
screen. So the program just automatically goes out of the loop and
returns back to the intro screen, which the program is logically doing.
I am dumb!lol